+++ lib/async_hooks.js
@@ -31,6 +31,7 @@ const { registerDestroyHook } = internal_async_hooks;
 const {
   executionAsyncId,
   triggerAsyncId,
+  getActiveResources,
   // Private API
   hasAsyncIdStack,
   getHookArrays,
@@ -347,6 +348,7 @@ module.exports = {
   createHook,
   executionAsyncId,
   triggerAsyncId,
+  getActiveResources,
   executionAsyncResource,
   // Embedder API
   AsyncResource,
