+++ lib/internal/async_hooks.js
@@ -1,10 +1,12 @@
 'use strict';
 
 const {
+  ArrayPrototypeForEach,
   ArrayPrototypeSlice,
   ErrorCaptureStackTrace,
   ObjectPrototypeHasOwnProperty,
   ObjectDefineProperty,
+  ObjectValues,
   Promise,
   Symbol,
 } = primordials;
@@ -544,9 +546,49 @@ function triggerAsyncId() {
 }
 
 
+const {
+  _getActiveHandles,
+  _getActiveRequests,
+} = internalBinding('process_methods');
+
+let internalTimers;
+
+function getActiveResources() {
+  if (internalTimers == null) {
+    internalTimers = require('internal/timers');
+  }
+  const handles = _getActiveHandles();
+  const reqs = _getActiveRequests();
+
+  const timers = {};
+  ArrayPrototypeForEach(ObjectValues(internalTimers.timerListMap), (list) => {
+    let timer = list._idlePrev === list ? null : list._idlePrev;
+
+    while (timer !== null) {
+      timers[timer[internalTimers.async_id_symbol]] = timer;
+
+      timer = timer._idlePrev === list ? null : list._idlePrev;
+    }
+  });
+
+  const immediates = {};
+  const queue = internalTimers.outstandingQueue.head !== null ?
+    internalTimers.outstandingQueue : internalTimers.immediateQueue;
+  let immediate = queue.head;
+  while (immediate !== null) {
+    immediates[immediate[internalTimers.async_id_symbol]] = immediate;
+
+    immediate = immediate._idleNext;
+  }
+
+  return { ...handles, ...reqs, ...timers, ...immediates };
+}
+
+
 module.exports = {
   executionAsyncId,
   triggerAsyncId,
+  getActiveResources,
   // Private API
   getHookArrays,
   symbols: {
