+++ lib/internal/util.js
@@ -3,7 +3,6 @@
 const {
   ArrayFrom,
   ArrayIsArray,
-  ArrayPrototypeForEach,
   ArrayPrototypePush,
   ArrayPrototypeSlice,
   ArrayPrototypeSort,
@@ -15,7 +14,6 @@ const {
   ObjectGetOwnPropertyDescriptors,
   ObjectGetPrototypeOf,
   ObjectSetPrototypeOf,
-  ObjectValues,
   Promise,
   ReflectApply,
   ReflectConstruct,
@@ -47,10 +45,6 @@ const {
   sleep: _sleep
 } = internalBinding('util');
 const { isNativeError } = internalBinding('types');
-const {
-  _getActiveHandles,
-  _getActiveRequests,
-} = internalBinding('process_methods');
 
 const noCrypto = !process.versions.openssl;
 
@@ -447,39 +441,6 @@ function createDeferredPromise() {
   return { promise, resolve, reject };
 }
 
-let internalTimers;
-
-function getActiveResources() {
-  if (internalTimers == null) {
-    internalTimers = require('internal/timers');
-  }
-  const handles = _getActiveHandles();
-  const reqs = _getActiveRequests();
-
-  const timers = {};
-  ArrayPrototypeForEach(ObjectValues(internalTimers.timerListMap), (list) => {
-    let timer = list._idlePrev === list ? null : list._idlePrev;
-
-    while (timer !== null) {
-      timers[timer[internalTimers.async_id_symbol]] = timer;
-
-      timer = timer._idlePrev === list ? null : list._idlePrev;
-    }
-  });
-
-  const immediates = {};
-  const queue = internalTimers.outstandingQueue.head !== null ?
-    internalTimers.outstandingQueue : internalTimers.immediateQueue;
-  let immediate = queue.head;
-  while (immediate !== null) {
-    immediates[immediate[internalTimers.async_id_symbol]] = immediate;
-
-    immediate = immediate._idleNext;
-  }
-
-  return { ...handles, ...reqs, ...timers, ...immediates };
-}
-
 module.exports = {
   assertCrypto,
   cachedResult,
@@ -490,7 +451,6 @@ module.exports = {
   deprecate,
   emitExperimentalWarning,
   filterDuplicateStrings,
-  getActiveResources,
   getConstructorOf,
   getSystemErrorMap,
   getSystemErrorName,
