+++ lib/util.js
@@ -67,7 +67,6 @@ const types = require('internal/util/types');
 
 const {
   deprecate,
-  getActiveResources,
   getSystemErrorMap,
   getSystemErrorName: internalErrorName,
   promisify
@@ -258,7 +257,6 @@ module.exports = {
   deprecate,
   format,
   formatWithOptions,
-  getActiveResources,
   getSystemErrorMap,
   getSystemErrorName,
   inherits,
