+++ test/parallel/test-async-hooks-getactiveresources-requests.js
@@ -3,7 +3,7 @@
 const common = require('../common');
 const assert = require('assert');
 const fs = require('fs');
-const { getActiveResources } = require('util');
+const { getActiveResources } = require('async_hooks');
 
 for (let i = 0; i < 12; i++)
   fs.open(__filename, 'r', common.mustCall());
